+++ Meizi/NetworkUtil.m
@@ -24,7 +24,7 @@ static NetworkUtil *singleton;
 }
 
 - (void)getMeizi:(NSString*)url
-           pages:(int)pages
+           pages:(NSInteger)pages
          success:(void (^)(NSString *succMsg,NSArray *meiziArray))success
          failure:(void (^)(NSString *failMsg,NSError *error))failure {
     NSNumber *page = [NSNumber numberWithInteger:pages];
